Fix bitrot in {default,shell}.nix
It was no longer possible to run the tests using
nix-shell as instructed in the README. There were
several issues:

  - default.nix had outdated dependencies
  - default.nix had `doCheck` set to false, which
    meant that the test dependencies (tasty, etc)
    wouldn't show up in the derivation and would
    not be installed.
  - haskell deps are to be found in `(propagated}buildInputs`
  - the pinned version of nixpkgs was very old, and
    prettyprinter was not available there
  - stdenv.lib was deprecated long ago and it is no longer
    available in recent versions of nixpkgs

Because of the last point, running nix-shell would effectively
only install ghc and cabal-install, and cabal would to build
llvm-hs, which is needed only for running the tests.

With the above fixed, running nix-shell would try to build
llvm-hs and llvm-hs-pure and fail: the default.nix files
there have also bitrot, have missing dependencies, etc.

Because llvm-hs and llvm-hs-pure are available in
nixpkgs now, we can rely on those versions instead. I'm
leaving the support for running versions from source, which
can be useful at some point once they are fixed upstream.